moved update quuantity method to Repository Items
